% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survHEinla-package.R
\docType{package}
\name{survHEinla-package}
\alias{survHEinla}
\alias{survHEinla-package}
\title{survHEinla: Survival Analysis in Health Economic Evaluation using INLA}
\description{
A module to complement the backbone structure of the package survHE and expand its functionality to run survival models under a Bayesian approach (based on Integrated Nested Laplace Approximation; the underlying 'INLA' package is available for download at \url{https://inla.r-inla-download.org/R/stable/}). \doi{10.18637/jss.v095.i14}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/giabaio/survHEinla}
  \item \url{https://gianluca.statistica.it/software/survhe}
  \item Report bugs at \url{https://github.com/giabaio/survHEinla/issues}
}

}
\author{
\strong{Maintainer}: Gianluca Baio \email{g.baio@ucl.ac.uk}

}
\keyword{internal}
